# docs/api-documentations/openapi/asset.openapi.json
{
  "components": {
    "schemas": {
...
      "RepairCompleteDto": {
        "properties": {
          "actualCost": {
            "description": "实际费用（元，无费用为 0；最多两位小数）",
            "example": "0",
            "type": "string"
          },
          "completedAt": {
            "description": "完成时间（RFC 3339；缺省为当前时间）",
            "type": "string"
          },
          "postStatus": {
            "description": "完成后资产恢复状态（使用中 / 闲置）",
            "enum": [
              "IN_USE",
              "IDLE"
            ],
            "type": "string"
          },
          "result": {
            "description": "维修结果",
            "maxLength": 1000,
            "type": "string"
          }
        },
        "required": [
          "result",
          "actualCost",
          "postStatus"
        ],
        "type": "object"
      },
      "RepairOrderCreateDto": {
        "properties": {
          "assetId": {
            "description": "资产 id",
            "minimum": 1,
            "type": "number"
          },
          "faultDescription": {
            "description": "故障/维修事项",
            "maxLength": 1000,
            "type": "string"
          },
          "idempotencyKey": {
            "description": "客户端为一次用户意图生成的幂等键；网络重试时保持不变",
            "maxLength": 128,
            "type": "string"
          },
          "reportedAt": {
            "description": "报修时间（RFC 3339；缺省为当前时间）",
            "type": "string"
          }
        },
        "required": [
          "assetId",
          "faultDescription"
        ],
        "type": "object"
      },
      "RepairStartDto": {
        "properties": {
          "startedAt": {
            "description": "开始维修时间（RFC 3339；缺省为当前时间）",
            "type": "string"
          }
        },
        "type": "object"
      },
      "StockChangeRequestCreateDto": {
        "properties": {
          "changedAt": {
            "description": "整单变更时间（RFC 3339；缺省提交时）",
            "type": "string"
          },
          "idempotencyKey": {
            "description": "客户端为一次用户意图生成的幂等键；网络重试时保持不变",
            "maxLength": 128,
            "type": "string"
          },
          "items": {
            "description": "变更明细行（同一库存条目整单只能一次）",
            "items": {
              "$ref": "#/components/schemas/StockChangeRequestItemDto"
            },
            "maxItems": 100,
            "type": "array"
          },
          "remark": {
            "description": "整单备注",
            "maxLength": 500,
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "StockChangeRequestItemDto": {
        "properties": {
          "changeTypeId": {
            "description": "变更类型（字典项 id；库存变更类型仅表示意外扣减原因）",
            "minimum": 1,
            "type": "number"
          },
          "inventoryItemId": {
            "description": "库存条目 id",
            "minimum": 1,
            "type": "number"
          },
          "qty": {
            "description": "扣减数量（正整数）",
            "minimum": 1,
            "type": "number"
          },
          "reason": {
            "description": "具体原因（必填）",
            "maxLength": 500,
            "type": "string"
          }
        },
        "required": [
          "inventoryItemId",
          "reason",
          "qty"
        ],
        "type": "object"
      },
      "StockInRequestCreateDto": {
        "properties": {
          "idempotencyKey": {
            "description": "客户端为一次用户意图生成的幂等键；网络重试时保持不变",
            "maxLength": 128,
            "type": "string"
          },
          "items": {
            "description": "入库明细行（同一品种+规格+库位只能一行）",
            "items": {
              "$ref": "#/components/schemas/StockInRequestItemDto"
            },
            "maxItems": 100,
            "type": "array"
          },
          "receivedAt": {
            "description": "整单申请时间（RFC 3339；缺省提交时；批准后作为批次入库时间）",
            "type": "string"
          },
          "remark": {
            "description": "整单备注",
            "maxLength": 500,
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "StockInRequestItemDto": {
        "properties": {
          "brandId": {
            "description": "品牌（字典项 id）",
            "minimum": 1,
            "type": "number"
          },
          "consumableId": {
            "description": "品种 id（必须启用）",
            "minimum": 1,
            "type": "number"
          },
          "qty": {
            "description": "数量（正整数）",
            "minimum": 1,
            "type": "number"
          },
          "spec": {
            "description": "规格",
            "maxLength": 100,
            "type": "string"
          },
          "supplierId": {
            "description": "供应商（字典项 id）",
            "minimum": 1,
            "type": "number"
          },
          "unitPrice": {
            "description": "单价（元，最多两位小数；可空）",
            "example": "10.00",
            "type": "string"
          },
          "warehouseId": {
            "description": "目标库位 id（必须启用）",
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "consumableId",
          "spec",
          "warehouseId",
          "qty"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "csrfHeader": {
        "description": "CSRF 双提交头（状态变更请求必须与 wbme_csrf Cookie 一致）",
        "in": "header",
        "name": "X-WBME-CSRF-Token",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "服务端会话 Cookie（HttpOnly；登录/激活成功后下发）",
        "in": "cookie",
        "name": "wbme_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "asset 部署单元（资产系统）OpenAPI 契约。构建期生成（主 PRD §9.5）：统一错误结构见 components.schemas.ErrorResponse；默认会话 Cookie + CSRF 头认证，公开路由无 security。",
    "title": "WBME asset API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/repair-orders": {
      "get": {
        "operationId": "RepairController_list",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "资产 id",
            "in": "query",
            "name": "assetId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "维修单状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "REPAIRING",
                "CANCELLED",
                "COMPLETED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Repair"
        ]
      },
      "post": {
        "operationId": "RepairController_register",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairOrderCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Repair"
        ]
      }
    },
    "/api/v1/repair-orders/{id}": {
      "get": {
        "operationId": "RepairController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Repair"
        ]
      }
    },
    "/api/v1/repair-orders/{id}/cancel": {
      "post": {
        "operationId": "RepairController_cancel",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Repair"
        ]
      }
    },
    "/api/v1/repair-orders/{id}/complete": {
      "post": {
        "operationId": "RepairController_complete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairCompleteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Repair"
        ]
      }
    },
    "/api/v1/repair-orders/{id}/start": {
      "post": {
        "operationId": "RepairController_start",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairStartDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Repair"
        ]
      }
    },
    "/api/v1/stock-change-requests": {
      "get": {
        "operationId": "StockChangeController_listHistory",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "审批状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "description": "发起人姓名",
            "in": "query",
            "name": "applicantName",
            "required": false,
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StockChange"
        ]
      },
      "post": {
        "operationId": "StockChangeController_submit",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockChangeRequestCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "StockChange"
        ]
      }
    },
    "/api/v1/stock-change-requests/mine": {
      "get": {
        "operationId": "StockChangeController_listMine",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "审批状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "description": "发起人姓名",
            "in": "query",
            "name": "applicantName",
            "required": false,
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StockChange"
        ]
      }
    },
    "/api/v1/stock-in-requests": {
      "get": {
        "operationId": "StockInController_listHistory",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "审批状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "description": "发起人姓名",
            "in": "query",
            "name": "applicantName",
            "required": false,
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StockIn"
        ]
      },
      "post": {
        "operationId": "StockInController_submit",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockInRequestCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "StockIn"
        ]
      }
    },
    "/api/v1/stock-in-requests/mine": {
      "get": {
        "operationId": "StockInController_listMine",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "审批状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "description": "发起人姓名",
            "in": "query",
            "name": "applicantName",
            "required": false,
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StockIn"
        ]
      }
    },
...
  },
  "security": [
    {
      "csrfHeader": [],
      "sessionCookie": []
    }
  ],
  "servers": [],
...
}